Mention possible effect of CRS on distances
After finishing the route calculation, a dialog box will show some
statistics on distances. The dialog will always show km as the unit for
these distances, even though the unit actually seems to be determined
by the unit of the Coordinate Reference System (CRS).

For example, OpenStreetMap uses WGS 84 (EPSG:4326), which QGIS lists as
a Geographic Coordinate System (GCS), which have an angular unit of
measure (degrees). To be able to measure meters, a Projected
Coordinate System needs to be used, so the users needs to [change the
CRS][1].

To point the user in the right direction, the following text is added
to the dialog box:

> (If the above values do not make sense, consider changing CRS.)

[1]: http://docs.qgis.org/2.8/en/docs/training_manual/vector_analysis/reproject_transform.html