Correct weight calculation per edge
When creating a graph before calculation, `build_graph()` adds an edge
for each individual segment (pair of nodes) of a line. Currently, the
weight given to each edge is the length of the entire line, which seems
wrong.

For example, consider a 400 meter street that is represented by a
polyline consisting of 5 nodes (4 segments). When creating the graph
for this, 4 edges will be added. When the length of the entire street
is used as the weight for each individual segment, the combined weight
of the street will be 4 * 400 = 1600.

This might lead to sub-optimal route calculation and wrong distances
being reported.

This patch corrects that by using the distance of the segment as the
weight for the graph edge created for it.

As far as I can tell, only the use of the plugin from QGIS was
affected by this, not the command-line use.